데미지 텍스트 소수점 버리고 정수형으로, HP바 텍스트 올림하고 정수형으로 변경

+++ Source/KeroroBattle/DamageTextWidget.cpp
@@ -49,8 +49,9 @@ void UDamageTextWidget::SetTextFromDamage(float Damage)
 {
 	if (DamageText)
 	{
+		
 		DamageText->SetColorAndOpacity(FSlateColor(FLinearColor(1.0f, 1.0f, 1.0f)));
-		DamageText->SetText(FText::AsNumber(Damage));
+		DamageText->SetText(FText::AsNumber(static_cast<int32>(Damage)));
 	}
 }
 
@@ -60,7 +61,7 @@ void UDamageTextWidget::SetTextFromCritDamage(float Damage)
 	{
 		DamageText->SetColorAndOpacity(FSlateColor(FLinearColor(1.0f, 0.05f, 0.0f)));
 		DamageText->SetRenderScale(FVector2D(1.3f));
-		DamageText->SetText(FText::AsNumber(Damage));
+		DamageText->SetText(FText::AsNumber(static_cast<int32>(Damage)));
 	}
 }
